requirements.txt
asgiref==3.2.7
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.3.0
mysqlclient==1.4.6
packaging==20.4
pluggy==0.13.1
py==1.8.1
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.4
wrapt==1.12.1
zipp==3.1.0
